Adding in enemy sprites spawn
I sat down for most of my time with you today. I have all the code, but
have not inputted it yet on how to correctly spawn enemies without them
randomly spawning on top of the player controlled sprite. Will be
adding this soon while it’s fresh in my head