Finished Enemy Collision and Added Game Over
Now, when the enemies collide the sprites change to blue and you can
try to make other enemies move into their dead bodies to kill them.
Also, added in two more enemies because the game was easy with three
enemies. Lastly, made an “L” pop up on the screen if the player loses.